requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.4.1
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
cffi==1.14.6
channels==3.0.4
channels-redis==3.2.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.0
gunicorn
psycopg2-binary
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-dotenv==1.4.2
django-js-asset==1.2.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.1
platformdirs==2.4.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pylint==2.6.2
pyOpenSSL==20.0.1
pytz==2021.1
regex==2021.7.6
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.2
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.4.0
